♻️ refactor #4 : Board/Cell 플래그 상태 변경 및 tsconfig path 설정
